Smart dynamic styling with CSS variables
We can make use of a CSS variable in the light DOM
and our shadow DOM component is able to pick it up.

By doing this, we are able to use that variable
into the shadow css and to provide a fallback as well.

So, we gain a lot of flexibility over the styling with
this approach for our component as ship it with some
styling and the same time we allow the client to change
it.