Observing Attribute Changes
By default JavaScript does not observe changes in all
attributes.

We have to inform to it which ones we want to observe.

We do that by doing return the attributes through the use
of a static get observedAttributes(). Which, in turn, return
an array with the selected attributes we want to observe
changes.

Then, whenever an attribute get changed the
attributeCallback() is fired back.