Using disconnectedCallback
As we have a function that gets executed when
the component is present in the real DOM, the
connectedCallback, we have another that gets
executed when the component is removed from
the DOM.

Therefore the disconnectedCallback is a lifecycle
method is the appropriate place to execute things
when the component is removed.

It is always a good idea to remove the eventListeners
attached to this component.

In this commit in addition to do doing that, we moved
declaration of the _tooltipIcon variable to the class
constructor just to have it there, to make it easier
to identify which variables are class variables and to
be able to reference it into another scopes.