Using disconnectedCallback
As we have a function that gets executed when
the component is present in the real DOM, the
connectedCallback, we have another that gets
executed when the component is removed from
the DOM.

Therefore the disconnectedCallback is a lifecycle
method is the appropriate place to execute things
when the component is removed.

It is always a good idea to remove the eventListeners
attached to this component.