Dispatching custom events
We are trying to listen to a cancel event from within the component
at the html file that uses it.

The way it is this event is trapped inside the component and therefore
can only be listened to from within it.

That is not ideal and it will be solved next.