Configuring Custom Events
In this commit, we show two ways of dispatching an
event.

One, where we dispatch the event from an element that
wraps the actual custom element, and therefore we have
to say that the event should bubble up until it finds
someone that handles it. And composed mean that the
event could leave the shadow DOM and eventually get
to the light DOM.

And another one where we dispatch the event through the
custom element itself. That way as we're already listening
the custom element, we don't need to pass this extra object.