Using @Prop() decorator
This decorator annotation brings the functionally that if we set or change this
attribute, or if we directly or programmatically set or change the title property
from outside JavaScript then Stencil will automatically change this property and
it will rerun the render() and will do so in a very efficiency manner so that is does
not rerender the entire DOM that was generated based on it, but only the parts of the
DOM that changed.

This is what the @Prop() decorator does, it adds an automated watcher you could say.