Understanding mutable Props
Props by default are immutable. You can't change by inside a component, you can change them by outside.
This is because Stencil has something called a unidirectional data flow, which means data
only flows in one direction, in this case from outside the component to inside the component.

If you want to change a Prop from inside a component, you can do so by adding another
configuration to this Prop, which is mutable to true.

There is another way of chaging data inside the component too, which
we'll see next.