Starting to build a tooltip again with Stencil
...this time.

In regard to sytling, pay attention on how it is said that the host
of the component should have a relative position, so as the next
element (underneath it) has a position absolute in regard to it.

Other than that, it is worth to remember the use of the @State()
decorator, which can be used to manage internal data for a component.

This means that a user cannot modify this data from outside the
component, but the component can modify it however it see fit.

Any changes to a @State() property will cause the components render()
to be called again.