Finishing the tooltip built with Stencil
Now, we just added a new property with @Prop() decorator.

Props are custom attribute/properties exposed publicly on the
element that developers can provide values for. So Props should
be used to pass data down from the parent to the child.

It is fundamental to know that a Prop is by default immutable
from inside the component logic.

However, it's possible to explicitly allow a Prop to be
mutated from inside the component, by declaring it as mutable
as a property of an object passed to it.

In some cases it may be useful to keep a Prop in sync with an
attribute. In this case you can use the reflectToAttr option
in the @Prop() decorator.