HTTP Request in components
In this commit it's worth to remind that we used the
builtin fetch API.

Also don't forget to bind the onFetchStockPrice located
into the form element, so as the method refers the context
of this to the component class instead of event that fired it.

Other than that, we used a @State() decorator which is ideal when
we want to manage internal data for a component and don't want that
the user modify this data from outside the component.