Accessing the Host element from TS class
Stencil components render their children declaratively in their
render method using JSX.

However, sometimes you will need to set properties on the host
element itself. Stencil provides a couple of ways to access and
update the host element.

In this commit we used one of them, the Element Decorator.

The @Element() decorator is how to get access to the host element
within the class instance.

This returns an instance of an HTMLElement, so we have access to all
methods available you have on HTML Element, because it is of the type
of a HTML Element.

That we got access to the component itself from within the class through
the use of a querySelector HTMLInputElement method. Note that the query
was done into the shadow DOM where the component lives.