Using references
In this commit we show an alternative to using
a querySelector into the shadow DOM from within
the component class, and use reference instead.

Reference is a feature of Stencil that executes
an arrow function for us and then we can assign
a class variable to this element.

I thought the use of reference simpler than
the querySelector, less code to type and easier
to understand.