Two way data binding and input validation
In this commit we assigned the value of the html input
element to a class variable achieving two way data
binding that way.

Also, we added another attribute to the element, that
sends an event whenever a user enters text a in a textbox.
This attribute is binded to a function that receives this
event and checks if the field has some 'valid' input for now.

Futher validations will be attached to this submit later.