Using the componentDidLoad() lifecycle hook
In this commit we used the componentDidLoad lifecycle hook
to check if a component has the attribute stockSymbol, and
if so, we fetch the stock price for this symbol.

Note that into HTML stockSymbol is not a valid attribute
name, or at least html is not case sensitive. So, we have
to use a dash between the two words and Stencil will automatically
convert this to the camelCase version.