Emitting custom events
In this commit, we created a new class property of type EventEmitter.

This property it's used to emit an event when a user clicks on one of the
list items, this event is then listened to by the web application for
now.

Next commit we'll make this event be listened directly from another
component if present.