Using the Listen decorator
In this commit we made an event emitted from one component
be listened in another component with the help of @Listen
decorator.

For that, as the event being listened to is not into the render
method of the component that is listening to it, we had to add
the body selector, to tell that the event could bubble to the
whole body element.