Using hostDate
Stencil components render their children declaratively in their
render method using JSX. However, sometimes you will need to set
properties on the host element itself. Stencil provides a couple of
ways to access and update the host element.

hostData() method

The hostData method can be implemented by any Stencil component and
provides a way to declaratively set properties on the host element
during rendering. The hostData method is called whenever render is,
so it is useful for reacing to state and prop changes.